requirements.txt
Flask>=2.2.0
pymongo>=4.4.0
python-dotenv>=1.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
